Catch up to SinglePhaseUpscaler::init() interface change
Commit OPM/opm-upscaling@886c463 removed the 'z_tol' parameter from
method SinglePhaseUpscaler::init(), preferring to retrieve the value
from the 'PINCH' keyword or defaulting to zero if 'PINCH' was not
specified.  This did not result in a build failure in module
opm-benchmarks due to additional, defaulted method parameters in the
init() method.  It did however lead to interpreting the 'ztol'
setting of "upscale_relperm_benchmark.cpp" as the (absolute) linear
residual tolerance.

That in turn, because ztol=0, resulted in the linear solver never
actually converging to an acceptable solution according to the
specified criteria and rather ending up taking the maximum number of
linear iterations (= system dimension), effectively increasing the
computational cost by a factor of about two.

This commit removes the 'ztol' parameter to restore the original
behaviour.